requirements.txt
# PM-OS Python Dependencies
# Install with: pip3 install -r requirements.txt

# Core
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.28.0

# Jira Integration
jira>=3.5.0

# Slack Integration
slack-sdk>=3.0.0

# GitHub Integration (optional - uses gh CLI primarily)
PyGithub>=1.59.0

# Google Workspace Integration
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# AI/LLM (optional - for deep research and analysis)
google-generativeai>=0.8.0
anthropic>=0.18.0
boto3>=1.34.0

# MCP Server (optional - requires Python 3.10+)
mcp>=0.1.0; python_version >= "3.10"

# Utilities
markdown>=3.4.0
python-dateutil>=2.8.0
